fix distcheck warnings.

+++ clamd/others.c
@@ -283,20 +283,20 @@ static int read_fd_data(struct fd_buf *buf)
       if (n < 0)
 	  return -1;
       if (msg.msg_flags & MSG_TRUNC) {
-	  logg("^Message truncated at %d bytes\n", n);
+	  logg("^Message truncated at %d bytes\n", (int)n);
 	  return -1;
       }
       if (msg.msg_flags & MSG_CTRUNC) {
 	  if (msg.msg_controllen > 0)
 	      logg("^Control message truncated at %d bytes, %d data read\n",
-		   msg.msg_controllen, n);
+		   (int)msg.msg_controllen, (int)n);
 	  else
 	      logg("^Control message truncated, no control data received, %d bytes read"
 #ifdef C_LINUX
 		   "(Is SELinux/AppArmor enabled, and blocking file descriptor passing?)"
 #endif
 		   "\n",
-		   n);
+		   (int)n);
 	  return -1;
       }
       if (msg.msg_controllen) {
